Thank you for applying tsx file and css module moved to Component folder
