+++ frontend/src/pages/index.ts
@@ -1,5 +1,5 @@
 import { Apply } from "./Apply.tsx";
 import { Candidates } from "./Candidates.tsx";
-import { ThankyouForApplying } from "./ThankyouForApplying.tsx";
+import { ThankyouForApplying } from "../components/ThankyouForApplying.tsx";
 
 export { Candidates, Apply, ThankyouForApplying };
